refactor: move CategoryFilter to data layer
