+++ app/src/main/java/com/pereyrarg11/cocktail/categories/ui/CategoriesScreenContent.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.pereyrarg11.cocktail.R
+import com.pereyrarg11.cocktail.categories.data.CategoryFilter
 import com.pereyrarg11.cocktail.categories.ui.alcohol.AlcoholCategoryRow
 import com.pereyrarg11.cocktail.categories.ui.model.CategoryItemDisplayable
 import com.pereyrarg11.cocktail.categories.ui.model.CategoryWrapperDisplayable
