+++ app/src/main/java/com/pereyrarg11/cocktail/categories/ui/alcohol/AlcoholCategoryCard.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.pereyrarg11.cocktail.R
-import com.pereyrarg11.cocktail.categories.ui.CategoryFilter
+import com.pereyrarg11.cocktail.categories.data.CategoryFilter
 import com.pereyrarg11.cocktail.categories.ui.model.CategoryItemDisplayable
 import com.pereyrarg11.cocktail.common.ui.CocktailPreview
