+++ app/src/main/java/com/pereyrarg11/cocktail/categories/ui/model/CategoryItemDisplayable.kt
@@ -1,7 +1,7 @@
 package com.pereyrarg11.cocktail.categories.ui.model
 
 import androidx.annotation.StringRes
-import com.pereyrarg11.cocktail.categories.ui.CategoryFilter
+import com.pereyrarg11.cocktail.categories.data.CategoryFilter
 
 data class CategoryItemDisplayable(
     @StringRes
